Redirect survey url to take survey path.
After posting updates to a questionnaire the url will be
a path that if it’s requested again via get will go through
unused surveyor display code that doesn’t work.

This puts a redirect in before that so requests will be sent
to the /surveys/:code/take url instead.