Copy `string_value` to `text_value` fields if response is text.
If the responses are from an old questionnaire where the answer was stored on the `string_value` field duplicated it to the `text_value` field.

This change was made to preserve urls that were longer than 255 chars as `string_value` is a varchar(255) and `text_value` is a text blob.