Generate separated general & legal yml documents for use in Transifex
Transifex/Rails I18n support needs the root of the translation file to
be the locale specifier so it can load it We want don’t want to have to
translate specific jurisdiction questions to all languages unnecessarily
so it will help if we keep these sections separate in the translation
interface.

This is just for generating the source documents, other translations
will be pulled from Transifex.