Hide response set tracking code from user
When pressing publish the response_set id is put
in the `flashes` so that the javascript is able to
log this as a publish event.

The template for show flashes was very generic and
would just render anything. This change makes it
skip showing a flash message if the key starts with
an underscore.

Fixes #1074