Stop updating the responses when continuing
Since the questionnaire autosaves and currently
updating the responses all at once can take
upwards of 15 minutes, remove the saving and just
run the prepublish checks instead.

This is slightly sad in that the questionnaire
can’t be updated without javascript but it doesn’t
work without javascript in other places currently
either.