Remove white boxes from homepage headings
The main heading and subheading on the homepage
had a solid white background, if the viewport was
slightly narrower than expected this sat over the
burst logo with a white box.

Changed the background to to transparent and added
a slight white text shadow so that heading was
readable when scrolled over.