Create campaign once before running
The code was previously referring to campaigns by
name so if more than one sidekiq job started at
once it could possibly create more than once
instance of the campaign.

Change it so that a `CertificationCampaign` object
is created before starting the jobs and store the
feed url that it uses on the object.