Show certification campaign results that have invalid URLs
Sometimes a certification campaign will be able to
answer all required questions to acquire a raw
certificate but the urls that describe resources
may be invalid or not resolve.

This causes the result to not be published.

Add a column that shows the overall validity of
URLs for the dataset.