Determine jurisdiction if campaign not made through web UI
If the campaign has been made through the API then
the campaign won’t have the jurisdiction. Try
to determine this for the moment.

This is a sign that the api isn’t well thought out
and should probably be changed. There will be less
use cases for campaigns from the api once the web
UI is being used. So it can be removed then.