Temporarily fix starting a questionnaire in another language.
The javascript in application.js:267 manually
creates a url badly so if there are parameters in
the url it will append .json on the end creating a
404. Which is how the javascript signals a problem
with the documentation_url†.

This will temporarily turn off the addition of
locale to URLs as it’s not needed until we
translate the rest of the website. The locale is
currently stored on the `ResponseSet` so can be
restored without needing a parameter passed
around.

† this all needs to be changed hence the temp fix.