Avoid deadlock in saving generator
Under load the update made when creating the
`response_set` for a `CertificateGenerator` is
deadlocking.

I think it’s because the dirty attributes tracking
isn’t working on relations and the `request`
column is a text blob, which means there is more
locking to go through for an update. This is just
a guess though.

This attempts to resolve it by reducing the number
of updates to that table. The `save!` in the
ensure block is to make sure the relation is saved
even if an exception aborts the generation.