Retry `save!` when a deadlock aborts the update
Two update queries in MySQL can deadlock if one
transaction gets the lock for the row and another
gets the lock for the index first (if it’s near by
in the index, i.e. close id numbers)

I tried shuffling the order the updates when into
the queue but that was not enough distribution to
solve this. The only solution is to retry the
query. Though ActiveRecord does not make this
easy.