Move Surveyor routes in file to make it clear they are independent
The Surveyor routes are mounted in independent engine, so where exactly they are placed doesn’t matter.

I’m not sure we should even be adding these are Surveyor routes. The the routes are handled by code in our app (not the Gem) I don’t think there is any benefit, and it is a bit confusing.