Get Surveyor's specs running as part of the main app
Also moved all the specs into the top-level `spec/` directory.

I couldn't get the controller spec working with the Engine routes. There is lots of advice online, some of it not applicable to our version of Rails and/or RSpec, and the rest doesn't seem to work. I have made those tests pending for now. If we move Surveyor out of the Engine they might become useful again.