Only load the locales we want to use
This makes sense now that we have the rails-i18n Gem because it would otherwise load lots of translations that are not needed.

It was actually causing routing to break when they were all getting loaded. I guessing this might be because of locales with more than 2 characters (en-GB, for example). Might cause a bit of trouble in the future, but I don't think it will be a big issue to solve if and when it arises.