Add rake task to scan for duplicate strings in translations
Some times duplicates are fine if there are places where text is likely or possible to change independently. But with an ever growing number of total translations I thought it worthwhile to have a tool to scan for these. Duplicates mean more work for translators, and more work when text that should be the same everywhere does change.