Removed the Privacy Impact Assessment question.
The question has already been translated before, but recently it was
reworded in the English version, so the localization did not apply
anymore and the question was showing up in English, even with the
Czech locale active.

We have consulted the question with a laywer and decided to delete
it for now, as it’s not really relevant for the Czech law system and
won’t be in the close future.

Fixes #16.