Fix broken test due to string/symbol mismatch
This test was never actually working properly. It has been checking for a string, but the real value was always a symbol.

The test changes to expect 'exemplar'. The assumtion here is that you meet a level if there are no questions for that level. I don't know if that is a correct assumption, but that's the way the code works. In any case it's not likely to happen on a real certificate.